descartes-content-engine/requirements.txt
anthropic>=0.40.0
feedparser>=6.0.11
beautifulsoup4>=4.12.0
requests>=2.31.0
pdfplumber>=0.10.0
python-dotenv>=1.0.0
jinja2>=3.1.0
schedule>=1.2.0
imaplib2>=3.6
lxml>=5.0.0
